// src/app/civil-time-recording-desktop/model/enum.ts
export enum RateType {
    Advice = 'Advice', // prep, attendance and advocacy
    Travel = 'Travel', // travel and waiting
    Counsel = 'Counsel', // Court Attendance / Conference with Counsel
    Letter = 'Letter',
    Letterin = 'LetterIn',
    Advocacy = 'Advocacy',
    Milage = 'Milage',
    FaresIncVAT = 'FaresIncVAT',
    Waiting = 'Prepration',
    EmailIn = 'EmailIn',
    EmailOut = 'EmailOut',
    CallIn = 'CallIn',
    CallOut = 'CallOut',
    LetterOut = 'LetterOut'
}